requirements.txt
setuptools>=65.5.0
wheel
gunicorn==21.2.0

flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0

numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
Werkzeug==2.3.7
